add persons/results into rankings type
